% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_contigs.R
\name{get_contigs}
\alias{get_contigs}
\title{Extract contigs}
\usage{
get_contigs(path, nodes)
}
\arguments{
\item{path}{a path (character vector) with order of nodes to be extracted.}
}
\value{
A reconstructed sentence (character vector)
}
\description{
Extracts contigs from nodes according to the provided path.
}
