% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph.R
\name{get_graph}
\alias{get_graph}
\title{Build directed graph}
\usage{
get_graph(nodes)
}
\arguments{
\item{nodes}{A vector of nodes (generated by get_nodes.R)}
}
\description{
Builds a de-Brujin-Graph via left-right k-1 overlaps
}
