% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kmers.R
\name{get_kmers}
\alias{get_kmers}
\title{Get kmers}
\usage{
get_kmers(sequence, k)
}
\arguments{
\item{sequence}{String to be subsetted}

\item{k}{Subsequence size, integer. must be smaller than sequence length}
}
\description{
A function to build k-mers from a sequence given a length k.
}
