% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodes.R
\name{get_nodes}
\alias{get_nodes}
\title{Get nodes}
\usage{
get_nodes(reads, k)
}
\arguments{
\item{reads}{A set of "reads" to be processed (A vector of strings)}

\item{k}{Subsequence length, integer. must be smaller than sequence length}
}
\description{
Generates nodes
}
