% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_a_walk.R
\name{walk_the_graph}
\alias{walk_the_graph}
\title{Take A Walk}
\usage{
walk_the_graph(graph)
}
\arguments{
\item{graph}{a directed graph (nested list) with nodes = names(list) and innodes/outnodes as sublists}
}
\description{
Walk through the provided graph
}
